init project react-ts
