--- assets/scss/components/_breakpoints.scss
@@ -1,107 +1,0 @@
-/*=============== BREAKPOINTS ===============*/
-/* For small devices */
-@media screen and (max-width: 320px){
-    .container{
-        margin-left: 1rem;
-        margin-right: 1rem;
-    }
-
-    .profile{
-        &__info{
-            column-gap: 1.5rem;
-        }
-        &__buttons{
-            flex-direction: column;
-            row-gap: 1rem;
-        }
-    }
-
-    .skills__box{
-        column-gap: 1rem;
-    }
-}
-
-/* For medium devices */
-@media screen and (min-width: 576px){
-    .projects__content{
-        grid-template-columns: 332px;
-        justify-content: center;
-    }
-    .filters__content{
-        width: 332px;
-        margin: 3rem auto;
-    }
-}
-
-@media screen and (min-width: 776px){
-    .projects__content,
-    .skills__content{
-        grid-template-columns: repeat(2, 332px);
-    }
-    .skills__content{
-        justify-content: center;
-        column-gap: 3rem;
-    }
-}
-
-/* For large devices */
-@media screen and (min-width: 992px){
-    .container{
-        margin-left: auto;
-        margin-right: auto;
-    }
-
-    .change-theme{
-        top: 2.5rem;
-        right: 2.5rem;
-    }
-
-    .profile{
-        padding-top: 4rem;
-
-        &__border{
-            width: 135px;
-            height: 135px;
-            margin-bottom: 1rem;
-        }
-        &__perfil{
-            width: 120px;
-            height: 120px;
-
-            & img{
-                width: 90px;
-            }
-        }
-        &__profession{
-            margin-bottom: 1.5rem;
-        }
-        &__info{
-            column-gap: 3rem;
-
-            &-description{
-                font-size: var(--small-font-size);
-            }
-        }
-        &__buttons{
-            column-gap: 2rem;
-        }
-    }
-
-    .projects{
-        &__content{
-            gap: 2rem 3rem;
-        }
-        &__modal{
-            padding: 1.5rem;
-        }
-    }
-
-    .skills__title{
-        font-size: var(--normal-font-size);
-        margin-bottom: 2.5rem;
-    }
-
-    .footer__copy{
-        margin: 4.5rem 0 2.5rem;
-    }
-}
