--- assets/scss/components/_button.scss
@@ -1,33 +1,0 @@
-/*=============== BUTTONS ===============*/
-.button{
-    display: inline-flex;
-    align-items: center;
-    column-gap: .5rem;
-    background-color: var(--first-color);
-    color: #fff;
-    padding: 1.15rem 1.5rem;
-    border-radius: .5rem;
-    transition: .3s;
-    box-shadow: 0 8px 24px hsla(var(--hue), var(--sat), var(--lig), .25);
-
-    & i{
-        font-size: 1.25rem;
-    }
-
-    &:hover{
-        background-color: var(--first-color-alt);
-    }
-
-    &__small{
-        padding: .75rem;
-        box-shadow: none;
-    }
-    &__gray{
-        background-color: var(--text-color-lighten);
-        color: var(--title-color);
-
-        &:hover{
-            background-color: var(--text-color-light);
-        }
-    }
-}
