--- assets/scss/components/_filters.scss
@@ -1,130 +1,0 @@
-/*=============== FILTERS TABS===============*/
-.filters{
-    &__content{
-        margin: 2rem 0 2.5rem;
-        background-color: var(--text-color-lighten);
-        padding: .375rem;
-        border-radius: .75rem;
-        display: flex;
-        justify-content: space-between;
-        column-gap: .5rem;
-    }
-    &__button{
-        width: 100%;
-        border: none;
-        outline: none;
-        padding: 1rem;
-        color: var(--title-color);
-        font-size: var(--small-font-size);
-        font-family: var(--body-font);
-        font-weight: 500;
-        border-radius: .75rem;
-        cursor: pointer;
-        background-color: transparent;
-        transition: .3s;
-
-        &:hover{
-            background-color: var(--body-color);
-        }
-    }
-}
-
-/*=============== PROJECTS ===============*/
-.projects{
-    &__card{
-        position: relative;
-        border-radius: 1rem;
-        overflow: hidden;
-
-        & img{
-            width: 100%;
-            height: 100%;
-        }
-
-        &:hover .projects__modal{
-            bottom: 0;
-        }
-    }
-    &__modal{
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        bottom: -100%;
-        left: 0;
-        background: linear-gradient(180deg,
-                    hsla(var(--hue), 24%, 40%, .3) 0%,
-                    hsla(var(--hue), 24%, 4%, 1) 95%);
-        display: grid;
-        align-items: flex-end;
-        padding: 1.5rem 1.25rem;
-        transition: .3s;
-    }
-    &__subtitle,
-    &__title{
-        color: #fff;
-    }
-    &__subtitle{
-        font-size: var(--smaller-font-size);
-    }
-    &__title{
-        font-size: var(--h3-font-size);
-        margin-bottom: .75rem;
-    }
-    &__button{
-        padding: .5rem;
-
-    }
-}
-
-/*=============== SKILLS ===============*/
-.skills{
-    &__content{
-        row-gap: 3.5rem;
-    }
-    &__title{
-        font-size: var(--h3-font-size);
-        text-align: center;
-        margin-bottom: 1.5rem;
-    }
-    &__box{
-        display: flex;
-        justify-content: center;
-        column-gap: 3rem;
-    }
-    &__group{
-        display: grid;
-        align-content: flex-start;
-        row-gap: 1rem;
-    }
-    &__data{
-        display: flex;
-        column-gap: .5rem;
-
-        & i{
-            font-size: 1rem;
-            color: var(--first-color);
-
-        }
-    }
-    &__name{
-        font-size: var(--normal-font-size);
-        font-weight: 500;
-        line-height: 18px;
-    }
-    &__level{
-        font-size: var(--smaller-font-size);
-    }
-}
-
-/* Hide and show projects & skills */
-.filters [data-content]{
-    display: none;
-}
-.filters__active[data-content]{
-    display: grid;
-}
-
-/* Activate button filter */
-.filter-tab-active{
-    background-color: var(--body-color);
-}
