--- assets/scss/components/_footer.scss
@@ -1,10 +1,0 @@
-/*=============== FOOTER ===============*/
-.footer{
-    &__copy{
-        display: block;
-        margin: 2.5rem 0 2rem;
-        text-align: center;
-        font-size: var(--smaller-font-size);
-        color: var(--text-color-light);
-    }
-}
