--- assets/scss/components/_profile.scss
@@ -1,92 +1,0 @@
-/*=============== PROFILE ===============*/
-.profile{
-    position: relative;
-    padding-top: 3.5rem;
-
-    &__container{
-        row-gap: 2rem;
-    }
-    &__data{
-        display: grid;
-        text-align: center;
-    }
-    &__perfil{
-        & img{
-            width: 75px;
-        }
-
-        width: 100px;
-        height: 100px;
-        background: linear-gradient(180deg,
-                    hsla(var(--hue), var(--sat), var(--lig), 1) 0%,
-                    hsla(var(--hue), var(--sat), var(--lig), .2) 100%);
-        border-radius: 50%;
-        overflow: hidden;
-        display: flex;
-        align-items: flex-end;
-        justify-content: center;
-    }
-    &__border{
-        border: 3.5px solid var(--first-color);
-        justify-self: center;
-        width: 115px;
-        height: 115px;
-        border-radius: 50%;
-        display: grid;
-        place-items: center;
-        margin-bottom: .75rem;
-    }
-    &__name{
-        font-size: var(--h2-font-size);
-    }
-    &__profession{
-        font-size: var(--smaller-font-size);
-        font-weight: 500;
-        color: var(--text-color);
-        margin-bottom: 1rem;
-    }
-    &__social{
-        display: flex;
-        justify-content: center;
-        column-gap: .75rem;
-
-        &-link{
-            font-size: 1.25rem;
-            color: var(--title-color);
-            transition: .3s;
-
-            &:hover{
-                color: var(--first-color);
-            }
-        }
-    }
-    &__info{
-        grid-template-columns: repeat(3, max-content);
-        justify-content: center;
-        column-gap: 2.5rem;
-
-        &-group{
-            text-align: center;
-        }
-        &-number{
-            font-size: var(--normal-font-size);
-            margin-bottom: .25rem;
-        }
-        &-description{
-            font-size: var(--smaller-font-size);
-            font-weight: 500;
-        }
-    }
-    &__buttons,
-    &__buttons-small{
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-    &__buttons{
-        column-gap: 1rem;
-    }
-    &__buttons-small{
-        column-gap: .25rem;
-    }
-}
