--- assets/scss/components/_scroll.scss
@@ -1,15 +1,0 @@
-/*=============== SCROLL BAR ===============*/
-::-webkit-scrollbar{
-    width: .6rem;
-    border-radius: .5rem;
-    background-color: hsl(var(--hue), 8%, 66%);
-    
-    &-thumb{
-        background-color: hsl(var(--hue), 8%, 54%);
-        border-radius: .5rem;
-        
-        &:hover{
-            background-color: hsl(var(--hue), 8%, 44%);
-        }
-    }
-}
