--- assets/scss/layout/_layout.scss
@@ -1,11 +1,0 @@
-/*=============== REUSABLE CSS CLASSES ===============*/
-.container{
-    max-width: 968px;
-    margin-left: 1.5rem;
-    margin-right: 1.5rem;
-}
-
-.grid{
-    display: grid;
-    gap: 1.5rem;
-}
