--- assets/scss/styles.scss
@@ -1,10 +1,0 @@
-@import 'config/variables';
-@import 'base/base';
-@import 'theme/theme';
-@import 'layout/layout';
-@import 'components/profile';
-@import 'components/button';
-@import 'components/filters';
-@import 'components/footer';
-@import 'components/scroll';
-@import 'components/breakpoints';
