+++ src/pages/App.tsx
@@ -1,0 +1,7 @@
+import Router from "../routes";
+
+const App = () => {
+  return Router;
+}
+
+export default App;
