+++ src/pages/Home/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () => {
+  return (
+    <h1 className='font-bold text-lg underline text-red-500'>Halo Huda</h1>
+  )
+}
+
+export default Home;
